+++ demo/chat/Main.hs
@@ -1,9 +1,10 @@
-import           Control.Concurrent.STM (newTChanIO)
 import           Control.Monad (when)
+import           Control.Monad.Logger (MonadLogger, runFileLoggingT)
 import           Data.Text (Text)
 import           RON.Store.Sqlite (runStore)
 import qualified RON.Store.Sqlite as Store
 import           Text.Pretty.Simple (pPrint)
+import           UnliftIO (MonadUnliftIO, liftIO, newTChanIO)
 
 import qualified Database
 import           Fork (forkLinked)
@@ -16,20 +17,21 @@ import           UI (initUI, runUI)
 
 main :: IO ()
 main = do
-  Options{dataDir, cmd} <- parseOptions
-  db <- Store.newHandle dataDir
-  case cmd of
-    Show -> Database.loadAllMessages db >>= pPrint
-    Post username text -> do
-      messageRef <-
-        runStore db $ Database.newMessage MessageContent{username, text}
-      putStrLn $ "created message: " <> show messageRef
-    RunNode nodeOptions -> runNode db nodeOptions
-    RunUI UIOptions{username} nodeOptions -> do
-      forkLinked $ runNode db nodeOptions
-      runUI' username db
+  Options{dataDir, cmd, logFile} <- parseOptions
+  runFileLoggingT logFile do
+    db <- Store.newHandle dataDir
+    case cmd of
+      Show -> Database.loadAllMessages db >>= pPrint
+      Post username text -> do
+        messageRef <-
+          runStore db $ Database.newMessage MessageContent{username, text}
+        liftIO $ putStrLn $ "created message: " <> show messageRef
+      RunNode nodeOptions -> runNode db nodeOptions
+      RunUI UIOptions{username} nodeOptions -> do
+        forkLinked $ runNode db nodeOptions
+        runUI' username db
 
-runUI' :: Text -> Store.Handle -> IO ()
+runUI' :: (MonadLogger m, MonadUnliftIO m) => Text -> Store.Handle -> m ()
 runUI' username db = do
   onMessagePosted      <- newTChanIO
   onMessageListUpdated <- newTChanIO
@@ -39,7 +41,9 @@ runUI' username db = do
   forkLinked $ Database.messagePoster onMessagePosted db
   runUI uiHandle
 
-runNode :: Store.Handle -> NodeOptions -> IO ()
+runNode ::
+  (MonadFail m, MonadLogger m, MonadUnliftIO m) =>
+  Store.Handle -> NodeOptions -> m ()
 runNode db options@NodeOptions{listenPorts, peers} = do
   when (null listenPorts && null peers) $
     fail